backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt>=5.3.1
ruff==0.11.12
phonenumbers==9.0.6
djangorestframework-bulk
pytest>=8.3.5
celery
django-filter
django-countries
django-timezone-field
Pillow>=11.2.1
pytest-cov